requirements.txt
alpaca-trade-api>=3.0.0
python-binance>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
ta>=0.10.0
python-dotenv>=1.0.0
schedule>=1.2.0
requests>=2.31.0
pytest>=7.0.0
